cleaning up the code a bit (to work propper like)
git-svn-id: https://groke.otago.ac.nz/ant/ant@526 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/instream_buffer.h
@@ -18,8 +18,8 @@ protected:
 
 protected:
 	unsigned char *buffer;
-	long position;
-	long position_of_end_of_buffer;
+	long long position;
+	long long position_of_end_of_buffer;
 
 public:
 	ANT_instream_buffer(ANT_memory *memory, ANT_instream *source);
